Clickhouse add restore.

+++ aliyun-java-sdk-clickhouse/src/main/java/com/aliyuncs/clickhouse/model/v20191111/DescribeBackupsResponse.java
@@ -85,7 +85,7 @@ public class DescribeBackupsResponse extends AcsResponse {
 
 		private String backupStartTime;
 
-		private Integer backupSize;
+		private Long backupSize;
 
 		private String backupEndTime;
 
@@ -127,11 +127,11 @@ public class DescribeBackupsResponse extends AcsResponse {
 			this.backupStartTime = backupStartTime;
 		}
 
-		public Integer getBackupSize() {
+		public Long getBackupSize() {
 			return this.backupSize;
 		}
 
-		public void setBackupSize(Integer backupSize) {
+		public void setBackupSize(Long backupSize) {
 			this.backupSize = backupSize;
 		}
